+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Migrated from deprecated ADAL to MSAL through the latest [azidentity SDK](https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/sdk/azidentity).
 - Added support for uploading blobs in cold and premium tier.
 - Support CPK for adls storage accounts.
+- Lazy-write support for async flush and close file call. Actual upload will be scheduled in background when this feature is enabled.
 
 ## 2.2.1 (2024-02-28)
 **Bug Fixes**